simplify the artifacts page, deleting the "ArtifactOverview" component
