+++ codegraph-viewer/frontend/app/components/page/main.tsx
@@ -25,5 +25,3 @@ export default class Main extends React.Component<{}, {}> {
         );
     }
 }
-
-export { Main };
