+++ codegraph-viewer/frontend/stub-server/data/artifacts/co.degraph_codegraph-core_1.0-RELEASE.json
@@ -1,0 +1,71 @@
+{
+    "name": "codegraph-core",
+    "organization": "co.degraph",
+    "version": "1.0-RELEASE",
+    "dependencies": [
+        {
+            "name": "commons-lang",
+            "organization": "commons-lang",
+            "version": "2.6",
+            "configurations": [
+                "compile"
+            ]
+        },
+        {
+            "name": "coollections",
+            "organization": "co.degraph",
+            "version": "0.0.1",
+            "configurations": [
+                "compile"
+            ]
+        },
+        {
+            "name": "spring-core",
+            "organization": "org.springframework",
+            "version": "4.2.7.RELEASE",
+            "configurations": [
+                "compile"
+            ]
+        },
+        {
+            "name": "spring-context",
+            "organization": "org.springframework",
+            "version": "4.2.7.RELEASE",
+            "configurations": [
+                "compile"
+            ]
+        },
+        {
+            "name": "log4j",
+            "organization": "log4j",
+            "version": "1.2.17",
+            "configurations": [
+                "compile"
+            ]
+        },
+        {
+            "name": "jdom",
+            "organization": "org.jdom",
+            "version": "2.0.2",
+            "configurations": [
+                "compile"
+            ]
+        },
+        {
+            "name": "junit",
+            "organization": "junit",
+            "version": "4.12",
+            "configurations": [
+                "testCompile"
+            ]
+        },
+        {
+            "name": "jaxen",
+            "organization": "jaxen",
+            "version": "1.1.6",
+            "configurations": [
+                "compile"
+            ]
+        }
+    ]
+}
